% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mnstockprops.r
\name{get_mnstockprops}
\alias{get_mnstockprops}
\title{Get probabilities of stock membership for humpback whales}
\usage{
get_mnstockprops(x, props.ind = FALSE, return.distr = FALSE)
}
\arguments{
\item{x}{Data frame containing four columns and one row per whale:
minimum and maximum latitudes of occurrence (two numeric columns), season of
occurrence (character, \code{"summer"} (Jun-Oct) or \code{"all"}) (all months),
and an assigned weight (e.g., proration value from serious injury determination
for an individual whale, or relative interaction risk in a stratum;
set to all ones if not relevant). Note that the function rounds all latitudes to
0.1 degrees precision.}

\item{props.ind}{Logical value indicating whether to return probabilities for
individual whales or for the total number of whales (must be \code{FALSE}
to get appropriate uncertainties for the total).}

\item{return.distr}{Logical value indicating whether to return full probability
distribution (2000 samples) for each estimated proportion (\code{TRUE}) or
only summary statistics (\code{FALSE}).}
}
\value{
A data frame with the following columns:
\item{id}{Only included if \code{props.ind = TRUE}: Sequential integers
corresponding to each input row (i.e., each whale).}
\item{minlat}{From input data; only included if \code{props.ind = TRUE}.}
\item{maxlat}{From input data; only included if \code{props.ind = TRUE}.}
\item{ssn}{From input data; only included if \code{props.ind = TRUE}.}
\item{w}{From input data; only included if \code{props.ind = TRUE}.}
\item{sim}{Simulation number. Only included if \code{return.distr = TRUE}.}
\item{stock}{Stock (using abbreviations provided in Details).}
\item{prop}{(Mean) proportion.}
\item{sd}{Standard deviation of proportion from simulation. Only included if
\code{return.distr = FALSE}.}
\item{q95l}{2.5th percentile of proportion from simulation. Only included if
\code{return.distr = FALSE}.}
\item{q80l}{10th percentile of proportion from simulation. Only included if
\code{return.distr = FALSE}.}
\item{q80u}{90th percentile of proportion from simulation. Only included if
\code{return.distr = FALSE}.}
\item{q95u}{97.5th percentile of proportion from simulation. Only included if
\code{return.distr = FALSE}.}
}
\description{
\code{get_mnstockprops} provides probabilities of membership in each of four
humpback whale stocks, given minimum and maximum latitude(s) along the U.S.
West Coast. It can provide probabilities for individual whales or spatial
strata, or averaged across multiple whales or weighted strata.
}
\details{
\code{get_mnstockprops} provides expected probabilities of stock membership
of humpback whales off the U.S. West Coast (as delineated under the U.S. MMPA),
given a range of potential latitudes. Probabilities can be returned at either
the individual level (for each set of minimum and maximum latitudes provided)
or as an overall average that can be applied to the total number of whales (or
across all strata provided). Uncertainty is derived from Monte Carlo simulations,
so uncertainty for mean probabilities across whales cannot be calculated post
hoc from statistics for individuals.

Estimated proportions are provided for four stocks (with abbreviations used in
the returned data frame):
\itemize{
\item Central America / Southern Mexico - California-Oregon-Washington (CASM)
\item Mainland Mexico - California-Oregon-Washington (MX-COW)
\item Hawai'i (HI)
\item Mexico-North Pacific (MX-NP)
}

The stock membership probabilities returned are based on stock proportions by
latitude, estimated for discrete latitude bins following the methods reported
in Curtis et al. (2025). Probabilities across multiple bins are calculated
as weighted averages using relative densities along the U.S. West Coast from
Becker et al. (2020) and in the Salish Sea from Wright et al. (2021). The
proportions and uncertainties applied here currently match those reported
in Curtis et al. (2025).

Minimum and maximum latitudes accepted by \code{get_mnstockprops} are 30.5 and
49, which map to the southernmost and northernmost extents of the U.S. West
Coast EEZ, including the Salish Sea. To limit proportions to the outer coast,
set maximum latitude less than or equal to 48.5. The Salish Sea can be
included by setting the maximum latitude to 49, and can be subset by setting
both minimum and maximum latitudes to 49 (note that proportions and relative
densities for the Salish Sea do include waters south of 48.5 too!). Latitude
limits between 48.5 and 49 will be rejected. Additional useful reference
latitudes: CA/OR border = 42, OR/WA border = 46.3.

Use of the following object names in your workspace while using this package
will disable its functionality and could lead to erroneous answers and unexpected
warnings or error messages: \code{lxi}, \code{lxi.mc}, and \code{mn.n.line}.
}
\examples{
whale.locs <- data.frame(minlat = c(34.5, 46), maxlat = c(40, 48.5),
                         ssn = c("all", "all"), w = c(1, 1))
get_mnstockprops(whale.locs)
get_mnstockprops(whale.locs, props.ind = TRUE)

}
\references{
Becker, E.A., Forney, K.A., Miller, D.L., Fiedler, P.C., Barlow, J., Moore,
J.E. 2020. Habitat-based density estimates for cetaceans in the California
Current Ecosystem based on 1991-2018 survey data. NOAA Technical Memorandum
NMFS-SWFSC-638.

Curtis, K. A., Calambokidis, J., and others. 2025. Harnessing the power of
photo-ID data for apportionment to migratory whale herds: U.S. West Coast
humpback whale stock proportions by latitude for the period 2019-2024. NOAA
Technical Memorandum NMFS-SWFSC-723.

Wright, B. M., Nichol, L. M., Doniol-Valcroze, T. 2021. Spatial density models
of cetaceans in the Canadian Pacific estimated from 2018 ship-based surveys.
DFO Can. Sci. Advis. Sec. Res. Doc. 2021/049. viii + 46 p. (Erratum: March 2022)
}
